Create a pyodide-py package (#909)
