+++ src/pyodide-py/pyodide/__init__.py
@@ -1,0 +1,6 @@
+from ._base import open_url, eval_code, find_imports, as_nested_list
+from .console import get_completions
+
+__version__ = "0.15.0"
+
+__all__ = ["open_url", "eval_code", "find_imports", "as_nested_list", "get_completions"]
